--- Src/WinChat.Infrastructure/Services/Contracts/Part.cs
@@ -1,6 +1,0 @@
-﻿namespace WinChat.Infrastructure.Services.Contracts;
-
-public class Part
-{
-    public string? Text { get; set; }
-}
