+++ Src/WinChat.Infrastructure/Services/Gemini/Contracts/Part.cs
@@ -1,0 +1,6 @@
+﻿namespace WinChat.Infrastructure.Services.Gemini.Contracts;
+
+public class Part
+{
+    public string? Text { get; set; }
+}
