+++ Src/WinChat.Infrastructure/Services/Gemini/Contracts/TextGenerationResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace WinChat.Infrastructure.Services.Contracts;
+﻿namespace WinChat.Infrastructure.Services.Gemini.Contracts;
 
 public class TextGenerationResponse
 {
