+++ Src/WinChat.Infrastructure/Services/Gemini/IGeminiService.cs
@@ -1,0 +1,8 @@
+﻿using WinChat.Infrastructure.Services.Gemini.Contracts;
+
+namespace WinChat.Infrastructure.Services.Gemini;
+
+public interface IGeminiService
+{
+    Task<TextGenerationResponse> GenerateText(TextGenerationRequest textGenerationRequest, CancellationToken cancellationToken);
+}
