+++ Src/WinChat.Infrastructure/Services/IApiTokenConfiguration.cs
@@ -1,0 +1,6 @@
+﻿namespace WinChat.Infrastructure.Services;
+
+public interface IApiTokenConfiguration
+{
+    Task SetApiToken(string apiToken);
+}
