--- Src/WinChat.Infrastructure/Services/IGenerateTextService.cs
@@ -1,9 +1,0 @@
-﻿using WinChat.Infrastructure.Services.Contracts;
-
-namespace WinChat.Infrastructure.Services;
-
-public interface IGenerateTextService
-{
-    Task<TextGenerationResponse> GenerateText(TextGenerationRequest textGenerationRequest, CancellationToken cancellationToken);
-    Task SetApiToken(string apiToken);
-}
